requirements.txt
Flask==3.0.0
flask-sock==0.7.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
requests
python-dotenv
twilio









